pyproject.toml
package-mode = false
[project]
name = "api-summarization"
version = "0.1.0"
description = "API for generating summaries from Wikipedia articles"
authors = [{name = "Lennon", email = "lennon.ribeiro@protonmail.com"}]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "sqlalchemy>=2.0.44,<3.0.0",
    "pydantic>=2.12.5,<3.0.0",
    "typing-inspect>=0.9.0,<0.10.0",
    "beautifulsoup4>=4.14.2,<5.0.0",
    "psycopg2-binary>=2.9.10,<3.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.32.0,<1.0.0",
    "alembic>=1.17.2,<2.0.0",
    "openai>=2.8.1,<3.0.0",
    "langchain>=0.3.0,<1.0.0",
    "langchain-openai>=0.2.0,<1.0.0",
    "langchain-text-splitters>=0.3.0,<1.0.0",
    "langchain-core>=0.3.0,<1.0.0",
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "httpx>=0.27.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "huggingface-hub>=0.26.0,<1.0.0",
    "fastapi-restful>=0.6.0,<1.0.0",
    "requests>=2.31.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
]

[tool.poetry]
packages = [{include = "app"}]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
